Move pipeline step manifests into modal app module
